Origin/andyy/morefunctions (#69)
* Update and add new functions

For functions that contain special characters as the input for the body before calling Invoke-WebRequest, the content gets converted to UTF8 encoding so they get stored correctly
New functions to create widgets that were not in the previous release, ability to create/update/remove project and self service portal components

* Update module version number to 0.0.40.0

* Update functions to pass script analyzer check

Added in ShouldProcess in functions with Set and New action verb. Remove Write-Host references.